Issue #252 extract out commonality between offheap and disk stores
